Rename both scripts
